pyproject.toml
[project]
name = "game-market"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.1",
    "fastapi[standard]>=0.115.8",
    "pydantic-settings>=2.7.1",
    "pydantic[email]>=2.10.6",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.6",
]

[tool.ruff.lint]
select = [
    # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "F", # Pyflakes
    # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "E", # pycodestyle
    "W", # Warning
    # https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    # https://docs.astral.sh/ruff/rules/#mccabe-c90
    "C", # Complexity (mccabe+) & comprehensions
    # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "UP", # pyupgrade
    # https://docs.astral.sh/ruff/rules/#isort-i
    "I", # isort
]
ignore = [
    "D", "TD",
    "ANN002", "ANN201", "ANN204",
    "FA102",  "INP001", "RUF002", "A002", "RET504", "RUF001"
]

[tool.ruff.format]
# Use `\n` line endings for all files
line-ending = "lf"

[tool.ruff.lint.per-file-ignores]
"src/alembic/*" = ["ALL"]   # alembic folder
